treerag/pyproject.toml
[project]
name = "treerag"
version = "0.1.0"
description = "TreeRAG - PageIndex style PDF retrieval system"
requires-python = ">=3.9"
dependencies = [
    "pdfplumber>=0.10.0",
    "pikepdf>=8.0.0",
    "openai>=1.12.0",
    "anthropic>=0.20.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
treerag = "main:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
